support calendar(schedule) for mts200 and fix #339
